Count both returned and fetched rows for scanners
The debug message emitted after closing a scanner now includes both the
number of rows returned from the generator function, and also the number
of rows actually fetched from the server. If scanners are not completely
iterated over (e.g. because of a 'break' statement in the for loop for
the scanner), these numbers may differ significantly. If this happens
often, this may be a hint that the batch_size is not optimal.