Add support for kerberized Thrift Servers
Since Apache Thrift 0.9.2 onwards includes a Python module for SASL,
we can use the TSaslClientTransport class to make secure connections.

This requires manual installation of "puresasl" and "kerberos" modules
to fully work. Since the "kerberos" module will only work in Python 2.7+,
this feature will not work for users of Python 2.6 or below.