Explicitly close the opened connection
If a connection pool is created in a multiprocess environment, e.g. in settings.py of a Django app hosted by uWSGI with multiple workers, spawning processes will inherit the open fd and may close a connection that the parent process is still using.

Unfortunately, I was unable to create a simple test case to trigger this.